+++ atlas/data_structure.py
@@ -287,6 +287,11 @@ SITES_SCHEMA = {
     'modified_by': {
         'type': 'string',
     },
+    'install': {
+        'type': 'boolean',
+        'default': True,
+        'nullable': True,
+    },
 }
 
 STATISTICS_SCHEMA = {