+++ atlas/instance_operations.py
@@ -416,14 +416,24 @@ def correct_fs_permissions(instance):
                     os.chown(file, -1, group.gr_gid)
 
 
-def sync_instances():
+def sync_instances(sid=None):
     """Copy the instance files to all of the relevant nodes.
+
+    Keyword Arguments:
+        sid {string} -- p1 sid for an instance (default: {None})
     """
-    log.info('Instances | Sync')
+
+    log.info('Instances | Sync | id - %s', sid)
     hosts = SERVERDEFS[ENVIRONMENT]['webservers'] + SERVERDEFS[ENVIRONMENT]['operations_server']
     # Sync INSTANCE_ROOT then WEB_ROOT
-    for root in [tuple([LOCAL_INSTANCE_ROOT, INSTANCE_ROOT]), tuple([LOCAL_WEB_ROOT, WEB_ROOT])]:
-        utilities.sync(root[0], hosts, root[1], exclude='opcache')
+    if sid:
+        sid_local_instance_root = LOCAL_INSTANCE_ROOT + '/' + sid
+        sid_instance_root = INSTANCE_ROOT + '/' + sid
+        utilities.sync(sid_local_instance_root, hosts, sid_instance_root, exclude='opcache')
+    else:
+        utilities.sync(LOCAL_INSTANCE_ROOT, hosts, INSTANCE_ROOT, exclude='opcache')
+
+    utilities.sync(LOCAL_WEB_ROOT, hosts, WEB_ROOT, exclude='opcache')
 
 
 def switch_web_root_symlinks(instance):
@@ -438,6 +448,7 @@ def switch_web_root_symlinks(instance):
     if instance['type'] == 'express':
         if instance['path'] != 'homepage':
             web_directory_path = '{0}/{1}'.format(LOCAL_WEB_ROOT, instance['path'])
+            web_directory_sid = '{0}/{1}'.format(LOCAL_WEB_ROOT, instance['sid'])
             # If the instance has a multipart path
             if "/" in instance['path']:
                 # Setup a base path, all items in 'path' except for the last one
@@ -452,9 +463,15 @@ def switch_web_root_symlinks(instance):
                 os.remove(web_directory_path)
             # If the instance is being taken down, change target for symlink
             if instance['status'] not in ['take_down', 'down']:
-                utilities.relative_symlink(instance_code_path_current, web_directory_path)
+                utilities.relative_symlink(instance_code_path_current, web_directory_sid)
+                if instance['path'] != instance['sid']:
+                    utilities.relative_symlink(
+                        instance_code_path_current, web_directory_path)
             elif instance['status'] in ['take_down', 'down']:
-                utilities.relative_symlink(SITE_DOWN_PATH, web_directory_path)
+                # Don't need to remove path symlink because it's handled above
+                log.debug('Instance | Web root symlinks | Remove symlink for sid after take down')
+                if os.path.islink(web_directory_sid):
+                    os.remove(web_directory_sid)
         elif instance['path'] == 'homepage':
             for link in CORE_WEB_ROOT_SYMLINKS:
                 source_path = "{0}/{1}".format(instance_code_path_current, link)