+++ atlas/tasks.py
@@ -380,7 +380,8 @@ def site_provision(site):
     # works properly.
     site['db_key'] = utilities.encrypt_string(utilities.mysql_password())
     # Set future site status for settings file creation.
-    site['status'] = 'available'
+    if site['status'] == 'pending':
+        site['status'] = 'available'
 
     try:
         log.debug('Site provision | Create database')
@@ -397,21 +398,23 @@ def site_provision(site):
     # Trigger rsync
     # ? Is a way to request a sync (w/ install chained)? Sync once when creating 5 instances
     log.info('Instance | Provision | Rsync')
-    instance_operations.sync_instances()
+    instance_operations.sync_instances(site['sid'])
     # Run install
-    try:
-        execute(fabric_tasks.site_install, site=site)
-    except Exception as error:
-        log.error('Site install failed | Error Message | %s', error)
-        raise
+    if site.get('install') and site['install'] is not False:
+        try:
+            execute(fabric_tasks.site_install, site=site)
+        except Exception as error:
+            log.error('Site install failed | Error Message | %s', error)
+            raise
     # Correct file permissions
     instance_operations.correct_fs_permissions(site)
-    instance_operations.sync_instances()
+    instance_operations.sync_instances(site['sid'])
 
     # Update instance record
-    patch_payload = {'status': 'available',
+    patch_payload = {'status': site['status'],
                      'db_key': site['db_key'],
-                     'statistics': site['statistics']}
+                     'statistics': site['statistics'],
+                     'install': None}
     patch = utilities.patch_eve('sites', site['_id'], patch_payload)
     log.debug('Site provision | Patch | %s', patch)
 
@@ -592,6 +595,10 @@ def site_update(site, updates, original):
                 site['status'] = 'installed'
                 instance_operations.switch_settings_files(site)
                 instance_operations.switch_web_root_symlinks(site)
+                statistics = utilities.get_single_eve('statistics', site['statistics'])
+                statistics_patch_payload = '{{"site": "{0}"}}'.format(site['_id'])
+                statistics_patch = utilities.patch_eve(
+                    'statistics', statistics['_id'], statistics_patch_payload)
                 deploy_update_database = True
                 patch_payload = '{"status": "installed"}'
                 deploy_drupal_cache_clear = True
@@ -611,7 +618,7 @@ def site_update(site, updates, original):
     # We want to run these commands in this specific order.
     log.info('Site Update | Closing operations commands | Sync - %s | PHP Cache clear - %s | Drush rr - %s; updb - %s ; cc - %s', sync_instances, deploy_php_cache_clear, deploy_registry_rebuild, deploy_update_database, deploy_drupal_cache_clear)
     if sync_instances:
-        instance_operations.sync_instances()
+        instance_operations.sync_instances(site['sid'])
         execute(fabric_tasks.clear_php_cache)
     if deploy_php_cache_clear:
         execute(fabric_tasks.clear_php_cache)
@@ -1276,7 +1283,7 @@ def update_settings_file(site, batch_id, count, total):
         log.error('Command | Update Settings file | Batch - %s | %s of %s | Instance - %s | Error - %s',
                   batch_id, count, total, site, error)
         raise
-    instance_operations.sync_instances()
+    instance_operations.sync_instances(site['sid'])
     execute(fabric_tasks.clear_php_cache)
 
 
@@ -1289,7 +1296,7 @@ def update_homepage_files():
     except Exception as error:
         log.error('Command | Update Homepage files | Error - %s', error)
         raise
-    instance_operations.sync_instances()
+    instance_operations.sync_instances(site['sid'])
 
 
 @celery.task
@@ -1344,19 +1351,24 @@ def correct_file_permissions(instance):
 
 @celery.task(time_limit=2000)
 def import_backup(env, backup_id, target_instance):
+    """Download and import a backup
+
+    Arguments:
+        env {[type]} -- [description]
+        backup_id {[type]} -- [description]
+        target_instance {[type]} -- [description]
     """
-    Download and import a backup
-    """
-    log.info('Import Backup | Source ENV - %s | Source Backup ID - %s | Target Instance - %s',
-             env, backup_id, target_instance)
+
+    log.info('Import Backup | Source ENV - %s | Source Backup ID - %s', env, backup_id)
     backup = requests.get(
         '{0}/backup/{1}'.format(API_URLS[env], backup_id), verify=SSL_VERIFICATION)
     log.info('Import Backup | Backup - %s', backup)
+
     target = utilities.get_single_eve('sites', target_instance)
     utilities.create_database(target['sid'], target['db_key'])
     instance_operations.instance_delete(target)
     instance_operations.instance_create(target)
-    instance_operations.instance_sync()
+    instance_operations.sync_instances()
     execute(fabric_tasks.import_backup, backup=backup.json(),
             target_instance=target, source_env=env)
     instance_operations.correct_fs_permissions(target)